Move OptionalNumericAssertions to FluentAssertions.Optional.Numeric namespace
